+++ ums-webservice-academic/src/main/java/org/ums/report/generator/teachersEvaluationReport/TesGenerator.java
@@ -15,6 +15,6 @@ public interface TesGenerator {
   void createTesReportSuperAdmin(String pDeptId, Integer pSemesterId, OutputStream pOutputStream) throws IOException,
       DocumentException;
 
-  void getQuestionWiseReports(String pDeptId, Integer pYear, Integer pSemester, Integer pSemesterId,
-      Integer pQuestionId, OutputStream pOutputStream) throws IOException, DocumentException;
+  void getQuestionWiseReports(String pDeptId, Integer pYear, Integer pSemester, Integer pSemesterId, Long pQuestionId,
+      OutputStream pOutputStream) throws IOException, DocumentException;
 }
